+++ src/components/cards/ItemCard.component.jsx
@@ -19,6 +19,8 @@ function ItemCard({
   handleUpdateItemQuantity,
   handleUpdateItem,
   visibleFields = DEFAULT_VISIBLE_FIELDS_SET,
+  handleArchiveItem,
+  handleUnarchiveItem,
 }) {
   const [isEditing, setIsEditing] = useState(false);
 
@@ -54,7 +56,11 @@ function ItemCard({
         ) : (
           <>
             <h2>{item.ItemName}</h2>
-
+            {handleUnarchiveItem && (
+              <button onClick={() => handleUnarchiveItem(item.id)}>
+                Unarchive
+              </button>
+            )}
             {handleUpdateItem && (
               <>
                 {fieldsToRender.map(({ key, label }) => (
@@ -63,6 +69,11 @@ function ItemCard({
                   </p>
                 ))}
                 <button onClick={() => setIsEditing(true)}>Edit</button>
+                {item.Status !== "archived" && item.Status !== "disabled" && (
+                  <button onClick={() => handleArchiveItem(item.id)}>
+                    Archive
+                  </button>
+                )}
               </>
             )}