+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [Unreleased]
+
+### Added
+
+- Add archive and unarchive functionality to ItemCard component
+
 ## [0.2.3]  - 2026-03-03
 
 ### Added